% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_domo_dataset.R
\name{append_domo_dataset}
\alias{append_domo_dataset}
\title{Create Dataset in Domo}
\usage{
append_domo_dataset(domo, df, ds_name, ds_description)
}
\arguments{
\item{domo}{The domo object created in set_up_domo}

\item{df}{The dataframe object that contains the import data}

\item{ds_name}{The name that the dataset will be called}

\item{ds_description}{An optional description of the dataset}
}
\value{
The id generated for the dataset created.
}
\description{
Creates a dataset in the domo instance using the parameters provided.
When called again, will continue to append the results to the created dataset.
}
\examples{
append_domo_dataset(domo, df1, 'Test', 'Generated for a Test')
append_domo_dataset(domo, df2, 'Test', 'Generated for a Test')
}
