% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_domo_dataset.R
\name{create_domo_dataset}
\alias{create_domo_dataset}
\title{Create Dataset in Domo}
\usage{
create_domo_dataset(domo, df, ds_name, ds_description = "")
}
\arguments{
\item{domo}{The domo object created in set_up_domo}

\item{df}{The dataframe object that contains the import data}

\item{ds_name}{The name that the dataset will be called}

\item{ds_description}{An optional description of the dataset}
}
\value{
The id generated for the dataset created.
}
\description{
Creates a dataset in the domo instance using the parameters provided
}
\examples{
create_domo_dataset(domo, df, 'Test', 'Generated for a Test')
}
