% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_domo_dataset.R
\name{get_domo_dataset}
\alias{get_domo_dataset}
\title{Get Domo Dataset}
\usage{
get_domo_dataset(domo, dsid)
}
\arguments{
\item{domo}{The domo object created in set_up_domo}

\item{dsid}{The dataset id copied from the Domo instance}
}
\value{
A dataframe object containing the rows and columns from the dataset id.
}
\description{
Used to retrieve datasets from Domo using a dataset id.
}
\examples{
get_domo_dataset(domo, 'c6ef4f5b-h408-4b7b-9792-1f62938e9e56')
}
