% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_domo.R
\name{set_up_domo}
\alias{set_up_domo}
\title{Set Up Domo}
\usage{
set_up_domo(client_id, client_secret)
}
\arguments{
\item{client_id}{The client id that for the domo instance}

\item{client_secret}{The client secret for the domo instance}
}
\value{
A domo class object that can be used to export and import from domo.
}
\description{
Creates a domo object that will be used to connect to a domo instance
}
\examples{
set_up_domo('client_id', 'client_secret'
}
