% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_domo_dataset.R
\name{update_domo_dataset}
\alias{update_domo_dataset}
\title{Update Dataset in Domo}
\usage{
update_domo_dataset(domo, df, dsid)
}
\arguments{
\item{domo}{The domo object created in set_up_domo}

\item{df}{The dataframe object that contains the import data}

\item{dsid}{The domo dataset id used to identify the dataset}
}
\description{
Updates a dataset in the domo instance using the parameters provided
}
\examples{
update_domo_dataset(domo, df, '226h46d5-b607-41b3-baf1-47bd669eb621')
}
